refactor: modularize page builder actions
