--- packages/ui/src/components/cms/page-builder/state/actions.js
@@ -1,1 +1,0 @@
-export {};
