+++ packages/ui/src/components/cms/page-builder/state/actions/setGridCols.ts
@@ -1,0 +1,9 @@
+import type { HistoryState } from "@acme/types";
+import type { SetGridColsAction } from "./types";
+
+export function setGridCols(
+  state: HistoryState,
+  action: SetGridColsAction,
+): HistoryState {
+  return { ...state, gridCols: action.gridCols };
+}
