+++ packages/ui/src/components/cms/page-builder/state/actions/update.ts
@@ -1,0 +1,8 @@
+import type { HistoryState } from "@acme/types";
+import type { UpdateAction } from "./types";
+import { commit } from "../history";
+import { updateComponent } from "./helpers";
+
+export function update(state: HistoryState, action: UpdateAction): HistoryState {
+  return commit(state, updateComponent(state.present, action.id, action.patch));
+}
